:recycle: refactor: [FE] Image 아톰 컴포넌트 수정
기존 ProfileImage 를 커스텀 로더를 사용하는 Image 컴포넌트로 리팩토링
