--- frontend/components/atoms/ProfileImage/index.tsx
@@ -1,23 +1,0 @@
-import type { NextPage } from 'next'
-import Image from 'next/image'
-
-interface profileProps {
-    src: string,
-    size: number,
-    alt?: string,
-}
-
-const profile: NextPage<profileProps> = ({ src, size, alt }) => {
-    return (
-        <Image
-            loader={() => src}
-            src={src}
-            alt={alt || 'Profile image'}
-            width={size}
-            height={size}
-            objectFit='contain'
-        />
-    );
-}
-
-export default profile
